Uses SSML for speech generation.
This unlocks "rate" and "style" control.

The "rate" and "style" controls are added to the Azure config, and to the
"override" pane. This way the they can be set both in "develper" and normal
mode, with "developer" mode being a bit more agile.

Finally, this commit also adds a "locale" filter to restrict to a small set of
locales that are usually used.